+++ engine/org.eclipse.birt.report.engine/src/org/eclipse/birt/report/engine/layout/emitter/IPageDevice.java
@@ -28,7 +28,7 @@ public interface IPageDevice
 	 *            background color.
 	 * @return the new page.
 	 */
-	IPage newPage( float width, float height, Color backgroundColor );
+	IPage newPage( int width, int height, Color backgroundColor );
 
 	/**
 	 * Closes this page device.
